+++ tinydb/database.py
@@ -15,6 +15,7 @@ class Document(dict):
     This is a transparent proxy for database records. It exists
     to provide a way to access a record's id via ``el.doc_id``.
     """
+
     def __init__(self, value, doc_id, **kwargs):
         super(Document, self).__init__(**kwargs)
 
@@ -76,6 +77,10 @@ class StorageProxy(object):
         self._storage = storage
         self._table_name = table_name
 
+    def _new_document(self, key, val):
+        doc_id = int(key)
+        return Document(val, doc_id)
+
     def read(self):
         raw_data = self._storage.read() or {}
 
@@ -89,8 +94,8 @@ class StorageProxy(object):
 
         docs = {}
         for key, val in iteritems(table):
-            doc_id = int(key)
-            docs[doc_id] = Document(val, doc_id)
+            doc = self._new_document(key, val)
+            docs[doc.doc_id] = doc
 
         return DataProxy(docs, raw_data)
 
@@ -138,7 +143,10 @@ class TinyDB(object):
         """
 
         storage = kwargs.pop('storage', self.DEFAULT_STORAGE)
-        table = kwargs.pop('default_table', self.DEFAULT_TABLE)
+        default_table = kwargs.pop('default_table', self.DEFAULT_TABLE)
+        self._cls_table = kwargs.pop('table_class', self.table_class)
+        self._cls_storage_proxy = kwargs.pop('storage_proxy_class',
+                                             self.storage_proxy_class)
 
         # Prepare the storage
         #: :type: Storage
@@ -149,7 +157,18 @@ class TinyDB(object):
         # Prepare the default table
 
         self._table_cache = {}
-        self._table = self.table(table)
+        self._table = self.table(default_table)
+
+    def __repr__(self):
+        args = [
+            'tables={}'.format(list(self.tables())),
+            'tables_count={}'.format(len(self.tables())),
+            'default_table_documents_count={}'.format(self.__len__()),
+            'all_tables_documents_count={}'.format(
+                ['{}={}'.format(table, len(self.table(table))) for table in self.tables()]),
+        ]
+
+        return '<{} {}>'.format(type(self).__name__, ', '.join(args))
 
     def table(self, name=DEFAULT_TABLE, **options):
         """
@@ -161,13 +180,14 @@ class TinyDB(object):
         :param name: The name of the table.
         :type name: str
         :param cache_size: How many query results to cache.
+        :param table_class: Which table class to use.
         """
 
         if name in self._table_cache:
             return self._table_cache[name]
 
-        table = self.table_class(StorageProxy(self._storage, name), name,
-                                 **options)
+        table_class = options.pop('table_class', self._cls_table)
+        table = table_class(self._cls_storage_proxy(self._storage, name), name, **options)
 
         self._table_cache[name] = table
 
@@ -265,6 +285,18 @@ class Table(object):
         self._query_cache = LRUCache(capacity=cache_size)
 
         data = self._read()
+        self._init_last_id(data)
+
+    def __repr__(self):
+        args = [
+            'name={!r}'.format(self.name),
+            'total={}'.format(self.__len__()),
+            'storage={}'.format(self._storage),
+        ]
+
+        return '<{} {}>'.format(type(self).__name__, ', '.join(args))
+
+    def _init_last_id(self, data):
         if data:
             self._last_id = max(i for i in data)
         else:
@@ -346,6 +378,11 @@ class Table(object):
 
         return current_id
 
+    def _get_doc_id(self, document):
+        if not isinstance(document, dict):
+            raise ValueError('Document is not a dictionary')
+        return self._get_next_id()
+
     def _read(self):
         """
         Reading access to the DB.
@@ -402,11 +439,7 @@ class Table(object):
         :returns: the inserted document's ID
         """
 
-        doc_id = self._get_next_id()
-
-        if not isinstance(document, dict):
-            raise ValueError('Document is not a dictionary')
-
+        doc_id = self._get_doc_id(document)
         data = self._read()
         data[doc_id] = document
         self._write(data)
@@ -425,7 +458,7 @@ class Table(object):
         data = self._read()
 
         for doc in documents:
-            doc_id = self._get_next_id()
+            doc_id = self._get_doc_id(doc)
             doc_ids.append(doc_id)
 
             data[doc_id] = doc
@@ -480,6 +513,41 @@ class Table(object):
                 cond, doc_ids
             )
 
+    def write_back(self, documents, doc_ids=None, eids=None):
+        """
+        Write back documents by doc_id
+
+        :param documents: a list of document to write back
+        :param doc_ids: a list of documents' ID which need to be wrote back
+        :returns: a list of documents' ID that have been written
+        """
+        doc_ids = _get_doc_ids(doc_ids, eids)
+
+        if doc_ids is not None and not len(documents) == len(doc_ids):
+            raise ValueError(
+                'The length of documents and doc_ids is not match.')
+
+        if doc_ids is None:
+            doc_ids = [doc.doc_id for doc in documents]
+
+        # Since this function will write docs back like inserting, to ensure
+        # here only process existing or removed instead of inserting new,
+        # raise error if doc_id exceeded the last.
+        if len(doc_ids) > 0 and max(doc_ids) > self._last_id:
+            raise IndexError(
+                'ID exceeds table length, use existing or removed doc_id.')
+
+        data = self._read()
+
+        # Document specified by ID
+        documents.reverse()
+        for doc_id in doc_ids:
+            data[doc_id] = documents.pop()
+
+        self._write(data)
+
+        return doc_ids
+
     def upsert(self, document, cond):
         """
         Update a document, if it exist - insert it otherwise.
@@ -495,7 +563,7 @@ class Table(object):
         if updated_docs:
             return updated_docs
         else:
-            return self.insert(document)
+            return [self.insert(document)]
 
     def purge(self):
         """
@@ -586,3 +654,6 @@ class Table(object):
 
 # Set the default table class
 TinyDB.table_class = Table
+
+# Set the default storage proxy class
+TinyDB.storage_proxy_class = StorageProxy